[Keras Batch Matmul] move functions to backend to match keras API
