+++ python/flexflow/keras/backend/__init__.py
@@ -16,6 +16,7 @@
 from __future__ import absolute_import
 from __future__ import print_function
 import sys
+from .backend_functions import batch_dot, BatchMatmul
 
 # Default backend: FlexFlow.
 _BACKEND = 'flexflow'
